pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "PyLibMS"
version = "2.0.0"
requires-python = ">=3.10"
description = "Python library built for the libMessageStudio (LMS) proprietary file formats from Nintendo. Supports MSBT, MSBP, and MSBF."
readme = "README.md"
authors = [{ name = "AbdyyEee" }]
keywords = ["nintendo", "msbf", "msbt", "msbp", "modding", "lms", "libmessagestudio"]
license = { file = "LICENSE" }
dependencies = ["PyYAML==6.0.1"]

[project.urls]
Documentation = "https://github.com/AbdyyEee/PylibMS/wiki"
Repository = "https://github.com/AbdyyEee/PylibMS"